requirements.txt
langchain
langchain-community
langchain-core
playwright
beautifulsoup4
python-dotenv
faiss-cpu
sentence-transformers
google-search-results
nest_asyncio
serpapi
langchain-openai